UBLE-31 rename: group route에 따른 파일 경로 수정
