+++ apps/user/src/app/(main)/home/components/PersonalSection.tsx
@@ -1,8 +1,19 @@
 import React from "react";
 import SectionHeader from "@/components/ui/SectionHeader";
 import DynamicCard from "@/components/ui/DynamicCard";
+import { BrandContent } from "@/types/brand";
 
 const PersonalSection = () => {
+  const testData: BrandContent = {
+    brandId: 104,
+    name: "스타벅스",
+    category: "푸드",
+    description: "커피를 팝니다.",
+    imgUrl: "",
+    isVIPcock: true,
+    minRank: "NONE",
+    bookmarked: false,
+  };
   return (
     <div className="space-y-4">
       <SectionHeader title="개인별 추천" />
@@ -10,7 +21,7 @@ const PersonalSection = () => {
         className="scrollbar-hide flex gap-4 overflow-x-auto px-4 pb-2"
         style={{ scrollbarWidth: "none", msOverflowStyle: "none" }}
       >
-        <DynamicCard />
+        <DynamicCard data={testData} />
       </div>
     </div>
   );
