+++ apps/user/src/app/(main)/mypage/history/components/BenefitList.tsx
