+++ apps/user/src/app/(main)/mypage/history/components/BenefitListItem.tsx
