--- apps/user/src/app/mypage/history/page.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import BenefitList from './components/BenefitList';
-
-const page = () => {
-  return (
-    <div className='min-h-screen bg-gray-50'>
-      <BenefitList />
-    </div>
-  );
-};
-
-export default page;
